add settings for code refactor suggestions
