+++ src/popup/pages/settings.tsx
@@ -4,9 +4,12 @@ import { goBack } from "react-chrome-extension-router";
 import Toggle from "../components/ToggleSwitch";
 import { useEffect, useState } from "react";
 
-export interface SettingsConfig {
-    aiPrDescription?: boolean;
-}
+export type SettingsConfig = Record<string, boolean | undefined>;
+
+const settingLabels: Record<string, string> = {
+    aiPrDescription: "AI PR Description",
+    codeRefactor: "Code Refactor",
+};
 
 const Settings = () => {
     const [settingsConfig, setSettingsConfig] = useState<SettingsConfig>({});
@@ -16,7 +19,7 @@ const Settings = () => {
             const settingsConfig = await chrome.storage.sync.get("osSettingsConfig");
 
             if (settingsConfig.osSettingsConfig === undefined) {
-                const defaultSettings = { aiPrDescription: false };
+                const defaultSettings = { aiPrDescription: true, codeRefactor: true };
 
                 await chrome.storage.sync.set({ osSettingsConfig: defaultSettings });
                 setSettingsConfig(defaultSettings);
@@ -53,12 +56,17 @@ const Settings = () => {
                 <main className="main-content text-white">
                     <h3 className="text font-medium text-base leading-10">Settings:</h3>
 
-                    { settingsConfig.aiPrDescription !== undefined &&
-            <Toggle
-                enabledSetting={settingsConfig.aiPrDescription}
-                settingLabel="AI PR Description"
-                settingName="aiPrDescription"
-            />}
+                    {
+                        Object.keys(settingsConfig).map(settingName => (
+                            <Toggle
+                                key={settingName}
+                                enabledSetting={settingsConfig[settingName]!}
+                                settingLabel={settingLabels[settingName]}
+                                settingName={settingName}
+                            />
+                        ))
+
+                    }
 
                     <div className="flex flex-col gap-2" />
                 </main>
