+++ src/utils/dom-utils/changeSuggestorButton.ts
@@ -1,10 +1,25 @@
 import { ChangeSuggestorButton } from "../../content-scripts/components/AICodeRefactor/ChangeSuggestorButton";
 import { GITHUB_REVIEW_SUGGESTION_SELECTOR } from "../../constants";
 import { isPublicRepository } from "../fetchGithubAPIData";
+import { SettingsConfig } from "../../popup/pages/settings";
 
 const injectChangeSuggestorButton = async (commentNode: HTMLElement) => {
     if (!(await isPublicRepository(window.location.href))) {
         return;
+    }
+
+    const settingsConfig = await new Promise(resolve => {
+        chrome.storage.sync.get("osSettingsConfig", result => {
+            resolve(result.osSettingsConfig);
+        });
+    });
+
+    if (settingsConfig) {
+        const { codeRefactor } = settingsConfig as SettingsConfig;
+
+        if (!codeRefactor) {
+            return;
+        }
     }
 
     const suggestChangesIcon = commentNode.getElementsByClassName(GITHUB_REVIEW_SUGGESTION_SELECTOR)[0];
